pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cap-gallery-tools"
version = "0.0.1"
dependencies = [
    "beautifulsoup4==4.11.1",
    "click==8.1.3",
    "Pillow==9.3.0",
    "requests==2.28.1"
]
